refactor(backend/copilot-bot): split handler into focused modules
handler.py (951 lines) becomes a ~380-line orchestrator; the streaming
machine + artifact delivery + thread rename move to turn_stream.py
(TurnStreamer), prompt/thread-name assembly to prompt.py, attachment upload
+ failure notes to attachments.py. TargetState converts to pydantic and
copilot/bot/handler.py comes OFF the dataclass known-offenders allowlist.
Pure move — no behavior change; test mock targets updated to the new paths.
