+++ autogpt_platform/backend/backend/copilot/bot/handler_test.py
@@ -22,13 +22,8 @@ from .adapters.base import (
     ReferencedConversation,
 )
 from .bot_backend import ChatTurnDeniedError, LinkTokenResult, ResolveResult
-from .handler import (
-    MAX_TURN_FILE_IDS,
-    MessageHandler,
-    TargetState,
-    build_thread_name,
-    clamp_thread_name,
-)
+from .handler import MAX_TURN_FILE_IDS, MessageHandler, TargetState
+from .prompt import build_thread_name, clamp_thread_name
 
 
 def _ctx(
@@ -121,7 +116,7 @@ def _capture_handler_tasks():
         return task
 
     with patch(
-        "backend.copilot.bot.handler.asyncio.create_task",
+        "backend.copilot.bot.turn_stream.asyncio.create_task",
         side_effect=_capturing,
     ):
         yield tasks
@@ -255,7 +250,7 @@ class TestResolveTarget:
             result = await handler._resolve_target(_ctx(), adapter)
         assert result == "thread-created"
         subscribe.assert_awaited_once_with("discord", "thread-created")
-        assert adapter.create_thread.await_args.args[2] == "AutoPilot: hello bot"
+        assert adapter.create_thread.await_args.args[2] == "AutoGPT: hello bot"
 
     @pytest.mark.asyncio
     async def test_channel_falls_back_to_parent_when_thread_creation_fails(self):
@@ -273,7 +268,7 @@ class TestThreadAdoption:
         # turn so the user gets an answer, but DON'T subscribe — future
         # messages here need another @ to wake it back up. This keeps the
         # bot from hijacking ongoing team conversations. The reply should
-        # carry the thread history so AutoPilot has context.
+        # carry the thread history so AutoGPT has context.
         handler = MessageHandler(_api())
         adapter = _adapter()
         enqueue = AsyncMock()
@@ -471,7 +466,7 @@ class TestBatching:
         adapter = _adapter()
 
         with patch(
-            "backend.copilot.bot.handler.get_redis_async",
+            "backend.copilot.bot.turn_stream.get_redis_async",
             new=AsyncMock(return_value=AsyncMock(get=AsyncMock(return_value=None))),
         ):
             await handler._stream_batch(
@@ -505,10 +500,12 @@ class TestBatching:
 
         with (
             patch(
-                "backend.copilot.bot.handler.get_redis_async",
+                "backend.copilot.bot.turn_stream.get_redis_async",
                 new=AsyncMock(return_value=AsyncMock(get=AsyncMock(return_value=None))),
             ),
-            patch("backend.copilot.bot.handler.Settings", return_value=fake_settings),
+            patch(
+                "backend.copilot.bot.turn_stream.Settings", return_value=fake_settings
+            ),
         ):
             await handler._stream_batch(
                 [("Bently", "u1", "hi")], _ctx(), adapter, "target-1"
@@ -551,10 +548,12 @@ class TestBatching:
 
         with (
             patch(
-                "backend.copilot.bot.handler.get_redis_async",
+                "backend.copilot.bot.turn_stream.get_redis_async",
                 new=AsyncMock(return_value=AsyncMock(get=AsyncMock(return_value=None))),
             ),
-            patch("backend.copilot.bot.handler.Settings", return_value=fake_settings),
+            patch(
+                "backend.copilot.bot.turn_stream.Settings", return_value=fake_settings
+            ),
         ):
             await handler._stream_batch(
                 [("Bently", "u1", "hi")], _ctx(), adapter, "target-1"
@@ -579,7 +578,7 @@ class TestBatching:
 
         with (
             patch(
-                "backend.copilot.bot.handler.get_redis_async",
+                "backend.copilot.bot.turn_stream.get_redis_async",
                 new=AsyncMock(return_value=redis),
             ),
             patch(
@@ -619,7 +618,7 @@ class TestBatching:
 
         with (
             patch(
-                "backend.copilot.bot.handler.get_redis_async",
+                "backend.copilot.bot.turn_stream.get_redis_async",
                 new=AsyncMock(return_value=redis),
             ),
             patch(
@@ -659,7 +658,7 @@ class TestBatching:
 
         with (
             patch(
-                "backend.copilot.bot.handler.get_redis_async",
+                "backend.copilot.bot.turn_stream.get_redis_async",
                 new=AsyncMock(return_value=redis),
             ),
             patch(
@@ -682,14 +681,14 @@ class TestBatching:
 
 class TestStreamFallback:
     """Covers the empty-response fallback, including the boundary-flush bug
-    where prior code posted 'AutoPilot didn't produce a response' even though
+    where prior code posted 'AutoGPT didn't produce a response' even though
     content had already been flushed mid-stream.
     """
 
     @staticmethod
     def _redis_patch():
         return patch(
-            "backend.copilot.bot.handler.get_redis_async",
+            "backend.copilot.bot.turn_stream.get_redis_async",
             new=AsyncMock(return_value=AsyncMock(get=AsyncMock(return_value=None))),
         )
 
@@ -767,13 +766,13 @@ class TestThreadNames:
     def test_build_thread_name_from_prompt(self):
         assert (
             build_thread_name("  tell me\nabout space  ", "Bently", 100)
-            == "AutoPilot: tell me about space"
+            == "AutoGPT: tell me about space"
         )
 
     def test_build_thread_name_truncates_to_platform_limit(self):
         name = build_thread_name("x" * 200, "Bently", 100)
         assert len(name) <= 100
-        assert name.startswith("AutoPilot: ")
+        assert name.startswith("AutoGPT: ")
         assert name.endswith("...")
 
     def test_build_thread_name_respects_a_larger_limit(self):
@@ -787,7 +786,7 @@ class TestThreadNames:
         )
 
     def test_clamp_thread_name_falls_back_when_blank(self):
-        assert clamp_thread_name("   ", 100) == "AutoPilot Chat"
+        assert clamp_thread_name("   ", 100) == "AutoGPT Chat"
 
     def test_clamp_thread_name_never_overruns_a_tiny_cap(self):
         # A tiny adapter cap must not make the slice index go negative.
@@ -813,7 +812,7 @@ class TestDeliverArtifact:
         )
         text = "Here is your result: [chart.png](workspace://abc#image/png)"
 
-        await handler._send_text_and_artifacts(
+        await handler._streamer._send_text_and_artifacts(
             adapter, "target-1", text, _ctx(), "sess-1"
         )
 
@@ -841,8 +840,10 @@ class TestDeliverArtifact:
         fake_settings.config.frontend_base_url = "https://app.example.com"
         fake_settings.config.platform_base_url = ""
 
-        with patch("backend.copilot.bot.handler.Settings", return_value=fake_settings):
-            await handler._send_text_and_artifacts(
+        with patch(
+            "backend.copilot.bot.turn_stream.Settings", return_value=fake_settings
+        ):
+            await handler._streamer._send_text_and_artifacts(
                 adapter,
                 "target-1",
                 "[big.zip](workspace://big)",
@@ -876,8 +877,10 @@ class TestDeliverArtifact:
         fake_settings.config.frontend_base_url = "https://app.example.com"
         fake_settings.config.platform_base_url = ""
 
-        with patch("backend.copilot.bot.handler.Settings", return_value=fake_settings):
-            await handler._send_text_and_artifacts(
+        with patch(
+            "backend.copilot.bot.turn_stream.Settings", return_value=fake_settings
+        ):
+            await handler._streamer._send_text_and_artifacts(
                 adapter, "target-1", "[x.png](workspace://x)", _ctx(), "sess-1"
             )
 
@@ -896,7 +899,7 @@ class TestDeliverArtifact:
         handler = MessageHandler(_api())
         adapter = _adapter()
 
-        await handler._send_text_and_artifacts(
+        await handler._streamer._send_text_and_artifacts(
             adapter, "target-1", "[x.png](workspace://x)", _ctx(), None
         )
 
@@ -975,7 +978,7 @@ class TestAttachments:
             ),
             patch("backend.copilot.bot.handler.sessions.set_session", new=AsyncMock()),
             patch(
-                "backend.copilot.bot.handler.get_redis_async",
+                "backend.copilot.bot.turn_stream.get_redis_async",
                 new=AsyncMock(return_value=redis),
             ),
         ):
@@ -1053,7 +1056,7 @@ class TestAttachments:
 
         await handler.handle(self._dm_ctx("look", [("a.png", "image/png")]), adapter)
 
-        # Files are NOT uploaded sessionless (where AutoPilot couldn't read
+        # Files are NOT uploaded sessionless (where AutoGPT couldn't read
         # them); the user is told they couldn't be attached and no file_ids
         # reach the turn.
         api.upload_workspace_files.assert_not_awaited()
@@ -1326,7 +1329,7 @@ class TestTurnDenied:
             raise ChatTurnDeniedError(
                 TurnDenial(
                     reason="paywalled",
-                    message="AutoPilot needs an active subscription.",
+                    message="AutoGPT needs an active subscription.",
                     button_label="Subscribe",
                     button_url="https://app.example/settings/billing",
                 )
@@ -1338,7 +1341,7 @@ class TestTurnDenied:
         adapter = _adapter()
 
         with patch(
-            "backend.copilot.bot.handler.get_redis_async",
+            "backend.copilot.bot.turn_stream.get_redis_async",
             new=AsyncMock(return_value=AsyncMock(get=AsyncMock(return_value=None))),
         ):
             await handler._stream_batch(
@@ -1365,7 +1368,7 @@ class TestTurnDenied:
         adapter = _adapter()
 
         with patch(
-            "backend.copilot.bot.handler.get_redis_async",
+            "backend.copilot.bot.turn_stream.get_redis_async",
             new=AsyncMock(return_value=AsyncMock(get=AsyncMock(return_value=None))),
         ):
             await handler._stream_batch(
